Update to use NSUInteger etc. Still have to update dragging code to use individual NSDraggingItems instead of one large NSImage.

+++ UKFinderIconCell/UKFinderIconCell.m
@@ -784,7 +784,7 @@
 // -----------------------------------------------------------------------------
 
 -(void) selectWithFrame:(NSRect)aRect inView:(NSView *)aView editor:(NSText *)textObj
-            delegate:(id)anObject start:(int)selStart length:(int)selLength
+            delegate:(id)anObject start:(NSInteger)selStart length:(NSInteger)selLength
 {
     NSRect textFrame, imageFrame;
